requirements.txt
fastmcp
requests
fastapi
uvicorn
langchain
langchain-google-genai
fastmcp
python-dotenv